docs + some renamings to make the docs more clear + moved minSize and maxSize from ImageValidation to UploadValidation
