--- src/Validation/CommonValidation.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Josegonzalez\Upload\Validation;
-
-use Josegonzalez\Upload\Validation\Traits\CommonValidationTrait;
-
-class CommonValidation
-{
-    use CommonValidationTrait;
-}
