+++ src/Validation/DefaultValidation.php
@@ -3,8 +3,10 @@
 namespace Josegonzalez\Upload\Validation;
 
 use Josegonzalez\Upload\Validation\Traits\UploadValidationTrait;
+use Josegonzalez\Upload\Validation\Traits\ImageValidationTrait;
 
-class UploadValidation
+class DefaultValidation
 {
     use UploadValidationTrait;
+    use ImageValidationTrait;
 }
