+++ src/Validation/Traits/ImageValidationTrait.php
@@ -71,36 +71,4 @@ trait ImageValidationTrait
         list(, $imgHeight) = getimagesize($check['tmp_name']);
         return $height > 0 && $imgHeight <= $height;
     }
-
-    /**
-     * Check that the file is above the minimum file upload size
-     *
-     * @param mixed $check Value to check
-     * @param int $size Minimum file size
-     * @return bool Success
-     */
-    public static function isAboveMinSize($check, $size)
-    {
-        // Non-file uploads also mean the size is too small
-        if (!isset($check['size']) || !strlen($check['size'])) {
-            return false;
-        }
-        return $check['size'] >= $size;
-    }
-
-    /**
-     * Check that the file is below the maximum file upload size
-     *
-     * @param mixed $check Value to check
-     * @param int $size Maximum file size
-     * @return bool Success
-     */
-    public static function isBelowMaxSize($check, $size)
-    {
-        // Non-file uploads also mean the size is too small
-        if (!isset($check['size']) || !strlen($check['size'])) {
-            return false;
-        }
-        return $check['size'] <= $size;
-    }
 }
